+++ doc/version_history.rst
@@ -6,3 +6,4 @@ v0.1.0
 ------
 
 * Migrate the codes from `ts_m2 <https://github.com/lsst-ts/ts_m2>`_.
+* Rename **Model** class to **Controller** class.